# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/categories": {
      "post": {
        "operationId": "CategoriesController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The category has been successfully created.",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "name": "Electrodomésticos",
                  "description": "Aparatos para el hogar que facilitan las tareas diarias, como refrigeradores, lavadoras o microondas."
                }
              }
            }
          },
          "400": {
            "description": "No fields provided to update",
            "content": {
              "application/json": {
                "example": {
                  "message": "No fields provided to update",
                  "error": "Bad Request",
                  "statusCode": 400
                }
              }
            }
          }
        },
        "summary": "Crear una nueva categoría",
        "tags": [
          "Categories"
        ]
      },
      "get": {
        "operationId": "CategoriesController_getCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de categorías obtenida exitosamente",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "name": "Electrodomésticos",
                    "description": "Aparatos para el hogar que facilitan las tareas diarias, como refrigeradores, lavadoras o microondas."
                  },
                  {
                    "id": 2,
                    "name": "Muebles",
                    "description": "Artículos para amueblar y decorar espacios, incluyendo mesas, sillas, sofás y camas."
                  },
                  {
                    "id": 3,
                    "name": "Ropa",
                    "description": "Prendas de vestir para diferentes estilos, climas y ocasiones."
                  }
                ]
              }
            }
          }
        },
        "summary": "Obtener todas las categorías",
        "tags": [
          "Categories"
        ]
      },
      "put": {
        "operationId": "CategoriesController_updateCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The category has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCategoryDto"
                },
                "example": {
                  "id": 1,
                  "name": "Electrodomésticos",
                  "description": "Aparatos para el hogar que facilitan las tareas diarias, como refrigeradores, lavadoras o microondas."
                }
              }
            }
          },
          "400": {
            "description": "Error with the request",
            "content": {
              "application/json": {
                "examples": {
                  "categoryIdRequired": {
                    "summary": "Category ID is required for update",
                    "value": {
                      "message": "Category ID is required for update.",
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  },
                  "categoryNotFound": {
                    "summary": "No category with that id found",
                    "value": {
                      "message": "Category with id 442 not found",
                      "error": "Not Found",
                      "statusCode": 404
                    }
                  },
                  "nofieldsToUpdate": {
                    "summary": "No fields provided to update",
                    "value": {
                      "message": "No fields provided to update",
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  },
                  "updateFailed": {
                    "summary": "Failed to update this field",
                    "value": {
                      "message": "Failed to update this field.",
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Actualizar una categoría existente",
        "tags": [
          "Categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_deleteCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The category has been successfully deleted.",
            "content": {
              "application/json": {
                "example": {
                  "success": "true"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "examples": {
                  "categoryIdRequired": {
                    "summary": "Category ID is required for update",
                    "value": {
                      "message": "Category ID is required for update.",
                      "error": "Bad Request",
                      "statusCode": 400
                    }
                  },
                  "categoryNotFound": {
                    "summary": "No category with that id found",
                    "value": {
                      "message": "Category with id 442 not found",
                      "error": "Not Found",
                      "statusCode": 404
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Eliminar una categoría existente",
        "tags": [
          "Categories"
        ]
      }
    },
...
    "/images/{folder}": {
      "post": {
        "operationId": "ImagesController_uploadFile",
        "parameters": [
          {
            "name": "folder",
            "required": true,
            "in": "path",
            "description": "Carpeta destino para subir el archivo",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Subir una imagen a una carpeta",
        "tags": [
          "Images"
        ]
      }
    },
    "/images/{folder}/{path}": {
      "put": {
        "operationId": "ImagesController_modifyFile",
        "parameters": [
          {
            "name": "folder",
            "required": true,
            "in": "path",
            "description": "Carpeta donde se encuentra la imagen",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "required": true,
            "in": "path",
            "description": "Nombre o ruta de la imagen a modificar",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Archivo de imagen con la nueva versión"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Modificar una imagen existente",
        "tags": [
          "Images"
        ]
      }
    },
    "/images": {
      "delete": {
        "operationId": "ImagesController_deleteFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "path": {
                    "type": "string",
                    "description": "Ruta del archivo a eliminar"
                  }
                },
                "required": [
                  "path"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Eliminar una imagen por su ruta",
        "tags": [
          "Images"
        ]
      }
    },
...
    "/configurations/{id}": {
      "get": {
        "operationId": "ConfigurationsController_getConfiguration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configuración obtenida correctamente.",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "title": "Terminos y condiciones",
                  "text": "1. Aceptación de los Términos\nAl acceder y utilizar la aplicación Red por la Ciberseguridad, usted acepta estar sujeto a estos términos y condiciones. Si no está de acuerdo con alguna parte de estos términos, no debe utilizar nuestra aplicación.\n\n2. Uso de la Aplicación\nEsta aplicación está diseñada para proporcionar información y herramientas relacionadas con la ciberseguridad. El usuario se compromete a:\n• Utilizar la aplicación de manera responsable y ética\n• No intentar comprometer la seguridad de la aplicación\n• Proporcionar información veraz y actualizada\n• Mantener la confidencialidad de sus credenciales de acceso\n\n3. Privacidad y Protección de Datos\nNos comprometemos a proteger su privacidad y datos personales de acuerdo con las leyes aplicables de protección de datos. La información recopilada se utiliza únicamente para los fines establecidos en nuestra política de privacidad.\n\n4. Responsabilidades del Usuario\nEl usuario es responsable de:\n• Mantener la seguridad de su cuenta\n• Utilizar contraseñas seguras\n• Reportar cualquier actividad sospechosa\n• Cumplir con las leyes locales aplicables"
                }
              }
            }
          }
        },
        "summary": "Obtiene la configuración por ID",
        "tags": [
          "Configurations"
        ]
      },
      "put": {
        "operationId": "ConfigurationsController_updateConfiguration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configuración actualizada correctamente.",
            "content": {
              "application/json": {
                "example": {
                  "result": "success"
                }
              }
            }
          }
        },
        "summary": "Actualiza la configuración por ID",
        "tags": [
          "Configurations"
        ]
      }
    }
  },
  "info": {
    "title": "Endpoints de CRUD Usuarios",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the category",
            "example": "Electronics"
          },
          "description": {
            "type": "string",
            "description": "The description of the category",
            "example": "Devices and gadgets"
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The ID of the category",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the category",
            "example": "Electronics"
          },
          "description": {
            "type": "string",
            "description": "The description of the category",
            "example": "Devices and gadgets"
          }
        },
        "required": [
          "id"
        ]
      },
      "DeleteCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The ID of the category to delete",
            "example": 1
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}